chore: minor reorgs of files in processors
