+++ apps/processing/common/src/main/kotlin/com/ethvm/common/config/ChainConfig.kt
@@ -3,10 +3,69 @@ package com.ethvm.common.config
 import com.ethvm.avro.processing.BalanceDeltaType
 import com.ethvm.avro.processing.TokenType
 import com.ethvm.common.extensions.ether
+import com.ethvm.common.extensions.milliEther
 import com.ethvm.db.tables.records.BalanceDeltaRecord
 import java.math.BigInteger
 import java.sql.Timestamp
 import java.util.Collections.emptyList
+
+data class ChainConstants(
+  val maxExtraDataSize: Int = 32,
+  val minGasLimit: Int = 125_000,
+  val gasLimitBoundDivisor: Int = 1024,
+  val minDifficutly: BigInteger = BigInteger.valueOf(131_072),
+  val difficultyBoundDivisor: BigInteger = BigInteger.valueOf(2048),
+  val expDifficultyPeriod: Int = 100_000,
+  val uncleGenerationLimit: Int = 7,
+  val uncleListLimit: Int = 2,
+  val bestNumberDiffLimit: Int = 100,
+  val blockReward: BigInteger = 1500.milliEther(),
+  // transaction signatures whose s-value is greater than secp256k1n/2 are considered invalid, introduced in Homestead release
+  val secp256k1n: BigInteger = BigInteger("fffffffffffffffffffffffffffffffebaaedce6af48a03bbfd25e8cd0364141", 16),
+  val longestChain: Int = 192,
+  val durationLimit: Int = 8,
+  val initialNonce: BigInteger = BigInteger.ZERO,
+  val maxContractSize: Int = Integer.MAX_VALUE,
+  // introduced in homestead
+  val createEmptyContractOnOutOfGas: Boolean = true,
+  // introduced in homestead, before homestead this opcode should generate an exception
+  val hasDelegateCallOpCode: Boolean = false
+) {
+
+  companion object {
+
+    val olympic = ChainConstants()
+
+    val frontier = ChainConstants(
+      durationLimit = 13,
+      blockReward = 5.ether(),
+      minGasLimit = 5_000
+    )
+
+    val homestead = frontier.copy(
+      createEmptyContractOnOutOfGas = false,
+      hasDelegateCallOpCode = true
+    )
+
+    val daoHardFork = homestead
+  }
+}
+
+enum class ChainId(val number: Int) {
+
+  Mainnet(1),
+  Ropsten(3),
+  Dev(17);
+
+  companion object {
+
+    fun forName(name: String) = values().firstOrNull { it.name.toLowerCase() == name.toLowerCase() }
+
+    fun forHex(hex: String) = forNumber(Integer.parseInt(hex.replace("0x", ""), 16))
+
+    fun forNumber(number: Int) = values().firstOrNull { it.number == number }
+  }
+}
 
 interface ChainConfig {
 
@@ -189,73 +248,4 @@ open class ConstantinopleConfig(parent: ChainConfig) : ByzantiumConfig(parent) {
 
 open class PetersburgConfig(parent: ChainConfig) : ConstantinopleConfig(parent) {
   override fun eip1283() = false
-}
-
-interface NetConfig {
-
-  val chainId: ChainId
-
-  val genesis: GenesisFile
-
-  fun chainConfigForBlock(number: BigInteger): ChainConfig
-
-  companion object {
-
-    val mainnet by lazy {
-      BaseNetConfig(
-        ChainId.Mainnet,
-        Genesis.Mainnet,
-        0L to FrontierConfig(),
-        1_150_000L to HomesteadConfig(),
-        1_920_000L to DaoHardForkConfig(),
-        2_463_000L to Eip150HardForkConfig(DaoHardForkConfig()),
-        2_675_000L to Eip160HardForkConfig(DaoHardForkConfig()),
-        4_370_000L to ByzantiumConfig(DaoHardForkConfig()),
-        7_280_000L to PetersburgConfig(DaoHardForkConfig())
-      )
-    }
-
-    val ropsten by lazy {
-      BaseNetConfig(
-        ChainId.Ropsten,
-        Genesis.Ropsten,
-        0L to HomesteadConfig(),
-        10L to RopstenConfig(HomesteadConfig()),
-        1_700_000L to RopstenConfig(ByzantiumConfig(DaoHardForkConfig())),
-        4_230_000L to RopstenConfig(ConstantinopleConfig(DaoHardForkConfig())),
-        4_939_394L to RopstenConfig(PetersburgConfig(DaoHardForkConfig()))
-      )
-    }
-
-    val dev by lazy {
-      BaseNetConfig(
-        ChainId.Dev,
-        Genesis.Dev,
-        0L to DevConfig(PetersburgConfig(DaoHardForkConfig()))
-      )
-    }
-
-    private val configByChainId = listOf(mainnet, ropsten, dev)
-      .map { it.chainId to it }
-      .toMap()
-
-    fun forChainId(chainId: ChainId) = configByChainId[chainId]
-  }
-}
-
-class BaseNetConfig(override val chainId: ChainId, genesis: Genesis, vararg configs: Pair<Long, ChainConfig>) :
-  NetConfig {
-
-  override val genesis: GenesisFile = genesis.load()
-
-  private val blockNumbers = configs.map { it.first.toBigInteger() }
-  private val chainConfigs = configs.map { it.second }
-
-  override fun chainConfigForBlock(number: BigInteger): ChainConfig {
-    var idx = 0
-    while ((idx < blockNumbers.size - 1) && number >= blockNumbers[idx]) {
-      idx += 1
-    }
-    return chainConfigs[idx]
-  }
 }
